Restruct React components. use copy & change trash implementation.
Add new compnent files.
Move many variables/functions in App.jsx to each components.
remove instead of rename for trash can.
