+++ src/DateList.jsx
@@ -1,13 +1,33 @@
+import { useEffect, useState } from "react";
+import { invoke } from "@tauri-apps/api/tauri";
+
 function DateList(props) {
+    const [dateList, setDateList] = useState([]);
+
+    useEffect((e) => {
+        getDates();
+        props.setReloadDates(false);
+    }, [props.reloadDates])
+
+    const getDates = () => {
+        invoke("get_dates").then((r) => {
+            let l = JSON.parse(r);
+            setDateList(l);
+        });
+    };
     return (
         <>
-            <p>List of Date <a href="#" onClick={() => props.getDates()}>⟳</a></p>
+            <p>List of Date <a href="#" onClick={() => getDates()}>⟳</a></p>
             <div className="dateList">
                 <ul>
-                    {props.dateList.map((l, i) => {
+                    {dateList.map((l, i) => {
                         let date = new Date(l.year + '/' + l.month + '/' + l.day).toLocaleString('default', { year: 'numeric', month: '2-digit', day: '2-digit' });
                         return (<li key={i} >
-                            <a href="#" onClick={(e) => props.getPhotos(e, undefined)} data-date={date} data-page={props.datePage[date]}>
+                            <a href="#" onClick={(e) => {
+                                props.setCurrentDate(date);
+                                props.toggleImporter(false);
+                            }
+                            } data-date={date} data-page={props.datePage[date]}>
                                 {date}
                             </a></li>)
                     })
