+++ src/PhotoDisplay.jsx
@@ -1,23 +1,94 @@
-import { convertFileSrc } from "@tauri-apps/api/tauri";
-import { useState } from "react";
+import { convertFileSrc, invoke } from "@tauri-apps/api/tauri";
+import { useEffect, useState } from "react";
 
 function PhotoDisplay(props) {
     const [dragPhotoInfo, setDragPhotoInfo] = useState([]);
     const [scrollLock, setScrollLock] = useState(false);
     const [photoDisplayImgClass, setPhotoDisplayImgClass] = useState("");
+    const [photoZoom, setPhotoZoom] = useState("100%");
+    const [photoZoomReady, setPhotoZoomReady] = useState(false);
+
+    const loaded = false;
+    useEffect((e) => {
+        console.log("loaded");
+        document.querySelector("#dummy-for-focus").focus({ focusVisible: true });
+        console.log(document.querySelector("#photoDisplay a"));
+        props.shortCutNavigation.onKeyDown = photoNavigation;
+        props.shortCutNavigation.onKeyUp = photoNavigationUp;
+        props.setShortCutNavigation(props.shortCutNavigation);
+    }, [loaded]);
+
 
     function dragPhotoStart(e) {
-        setPhotoDisplayImgClass("photo_dragging"); photos
+        setPhotoDisplayImgClass("photo_dragging");
         setDragPhotoInfo({ is_dragging: true, x: e.clientX, y: e.clientY });
+    }
+
+    function photoNavigation(e) {
+        console.log(e);
+        let f = props.currentPhotoPath;
+        if (e.keyCode === 39) { // right arrow
+            nextPhoto(f)
+        } else if (e.keyCode === 37) { // left arrow
+            prevPhoto(f)
+        } else if (e.keyCode === 46) { // Del
+            moveToTrashCan(f)
+        } else if (e.ctrlKey && e.keyCode === 48) { // ctrl+0
+            setPhotoZoom("100%");
+        } else if (e.ctrlKey) {
+            console.log("ready");
+            setPhotoZoomReady(true);
+        }
+    }
+    function photoNavigationUp(e) {
+        console.log("release");
+        if (e.ctrlKey) {
+            console.log("release");
+            setPhotoZoomReady(false);
+        }
+    }
+
+    async function prevPhoto(f) {
+        await invoke("get_prev_photo", { path: f, dateStr: props.currentDate, sortValue: parseInt(props.sortOfPhotos) }).then((r) => {
+            if (r !== "") {
+                setPhotoZoom("100%");
+                props.setCurrentPhotoPath(r);
+            }
+        });
+    }
+
+    async function nextPhoto(f) {
+        await invoke("get_next_photo", { path: f, dateStr: props.currentDate, sortValue: parseInt(props.sortOfPhotos) }).then((r) => {
+            if (r !== "") {
+                setPhotoZoom("100%");
+                props.setCurrentPhotoPath(r);
+            }
+        });
+    }
+
+    function moveToTrashCan(f) {
+        console.log("delete file: " + f)
+        invoke("move_to_trash", { dateStr: props.currentDate, pathStr: f, sortValue: parseInt(props.sortOfPhotos) }).then((r) => {
+            console.log("target:", r);
+            if (!r) {
+                props.setCurrentDate("");
+                getPhotos(undefined, undefined);
+                props.setCurrentPhotoPath("");
+            } else {
+                props.setCurrentPhotoPath(r);
+                console.log("select photo:", r);
+                props.setCurrentPhotoPath(r);
+            }
+        });
     }
 
     // TODO: not correct scroll adjustment.
     function photoScroll(e) {
-        if (scrollLock || !props.zoomReady) {
+        if (scrollLock || !photoZoomReady) {
             return;
         }
         setScrollLock(true);
-        let zoom = parseInt(props.zoom.replace("%", ""));
+        let zoom = parseInt(photoZoom.replace("%", ""));
 
         const imgTag = document.querySelector(".photo img");
         const display = e.currentTarget.parentElement;
@@ -37,7 +108,8 @@ function PhotoDisplay(props) {
             zoom += 10;
         }
 
-        props.setZoom(zoom + "%");
+        setPhotoZoom(zoom + "%");
+        console.log(zoom);
 
         const sTop = (imgTag.height * yPos - display.clientHeight * yPos);
         const sLeft = (imgTag.width * xPos - display.clientWidth * xPos);
@@ -64,20 +136,27 @@ function PhotoDisplay(props) {
         }
     }
 
+    function closePhotoDisplay() {
+        props.setShowPhotoDisplay(false);
+    }
+
     function dragPhotoEnd(e) {
         setPhotoDisplayImgClass("");
         setDragPhotoInfo({});
     }
 
     return (
-        <div id="photoDisplay" autoFocus={true} className={props.class}>
+        <div className="photoDisplay" id="photoDisplay" autoFocus={true}>
             <p>Photo Viewer</p>
-            <a href="#" onClick={() => props.prevPhoto(props.path)}>&lt;&lt; prev</a>&nbsp;&nbsp;
-            || <a href="#" onClick={() => props.toggleCenterDisplay()}>close</a> ||&nbsp;&nbsp;
-            <a href="#" onClick={() => props.nextPhoto(props.path)}>next &gt;&gt;</a><br /><br />
-            <a href="#" onClick={() => props.moveToTrashCan(props.path)}>&#128465;</a>
+            <a href="#" id="dummy-for-focus">{/* Dummy */}</a>
+            <a href="#" onClick={() => prevPhoto(props.currentPhotoPath)}>&lt;&lt; prev</a>&nbsp;&nbsp;
+            || <a href="#" onClick={() => closePhotoDisplay()}>close</a> ||&nbsp;&nbsp;
+            <a href="#" onClick={() => nextPhoto(props.currentPhotoPath)}>next &gt;&gt;</a><br /><br />
+            <a href="#" onClick={() => moveToTrashCan(props.currentPhotoPath)}>&#128465;</a>
             <div className="photo">
-                <img className={photoDisplayImgClass} src={convertFileSrc(props.path)} width={props.zoom}
+                <img className={photoDisplayImgClass}
+                    src={convertFileSrc(props.currentPhotoPath)}
+                    width={photoZoom}
                     onMouseDown={(e) => dragPhotoStart(e)}
                     onMouseMove={(e) => dragPhoto(e)}
                     onMouseUp={(e) => dragPhotoEnd(e)}
