+++ src/PhotoLoading.jsx
@@ -1,0 +1,10 @@
+function PhotoLoading() {
+    return (
+        <div id="photoLoading" className="photoLoadingOn">
+            <h1>Now Loading Photos ...</h1>
+            <div className="lds-dual-ring"></div>
+        </div>
+    )
+}
+
+export default PhotoLoading;
